+++ samplegaussian.hh
@@ -1,15 +1,15 @@
 class SampleGaussian
 {
-  // Class to generate random numbers (float) with Gaussian distribution
+  // Class to generate random numbers (double) with Gaussian distribution
   // Box-Muller algorithm from Numerical Recipes
 public:
   SampleGaussian() : gotone(false) {}
 
   // Get number with zero mean & unit variance
-  float Get();
-  float Get(const float& Mean, const float& SD);
+  double Get();
+  double Get(const double& Mean, const double& SD);
 
 private:
-  float previous; // algorithm generates numbers in pairs, 
+  double previous; // algorithm generates numbers in pairs, 
   bool gotone;    //  so buffer one of them
 };
